requirements.txt
backoff==2.2.1
cachetools==5.3.0
colorama==0.4.5
openai==0.27.2
pymongo==4.3.3
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
redis==4.5.4
requests==2.28.2
telegram==0.0.1
